Break out Download and RequestManager into its own file
