+++ UWP/CommonUWP/CommonUWP.vcxproj.filters
@@ -444,6 +444,9 @@
     <ClCompile Include="..\..\Common\System\OSD.cpp">
       <Filter>System</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\Common\Net\HTTPRequest.cpp">
+      <Filter>Net</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="targetver.h" />
@@ -840,6 +843,9 @@
     </ClInclude>
     <ClInclude Include="..\..\Common\System\OSD.h">
       <Filter>System</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\Common\Net\HTTPRequest.h">
+      <Filter>Net</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
